src/main/java/models/Hero.java
package models;

public class Hero {

    private String name;
    private int age;
    private String power;
    private String weakness;
    

public Hero(String name, Integer age, String power ,String weakness) {
    this.name = name;
    this.age =age;
    this.power = power;
    this.weakness = weakness;
}

    public String getName() {return this.name;}
    public int getAge(){return this.age;}
    public String getPower(){return this.power;}
    public String getWeakness(){return this.weakness;}

}
